refactor(api): move repositories to domain folders
