+++ api/src/prescription/organization-learner/domain/usecases/index.js
@@ -11,7 +11,7 @@ import { userToCreateRepository } from '../../../../identity-access-management/i
 import * as organizationFeaturesAPI from '../../../../organizational-entities/application/api/organization-features-api.js';
 import { tagRepository } from '../../../../organizational-entities/infrastructure/repositories/tag.repository.js';
 import * as libOrganizationLearnerRepository from '../../../../prescription/organization-learner/infrastructure/repositories/organization-learner-repository.js';
-import * as combinedCourseRepository from '../../../../quest/infrastructure/repositories/combined-course-repository.js';
+import * as combinedCourseRepository from '../../../../quest/infrastructure/repositories/combined-courses/combined-course-repository.js';
 import * as questRepository from '../../../../quest/infrastructure/repositories/quest-repository.js';
 import { cryptoService } from '../../../../shared/domain/services/crypto-service.js';
 import { tokenService } from '../../../../shared/domain/services/token-service.js';
