+++ api/src/quest/infrastructure/repositories/combined-course-participations/organization-learner-participation-repository.js
@@ -1,8 +1,8 @@
-import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
+import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 import {
   OrganizationLearnerParticipation,
   OrganizationLearnerParticipationTypes,
-} from '../../domain/models/combined-course-participations/entities/OrganizationLearnerParticipation.js';
+} from '../../../domain/models/combined-course-participations/entities/OrganizationLearnerParticipation.js';
 
 export const findByOrganizationLearnerIdAndModuleIds = async ({ organizationLearnerId, moduleIds }) => {
   const participationById = await findByOrganizationLearnerIdsAndModuleIds({
