+++ api/src/quest/infrastructure/repositories/course-repository.js
@@ -3,7 +3,7 @@ import * as areaRepository from '../../../shared/infrastructure/repositories/are
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
 import * as tubeRepository from '../../../shared/infrastructure/repositories/tube-repository.js';
 import { COURSE_ITEM_TYPES, CourseItem } from '../../domain/models/combined-courses/value-objects/CourseItem.js';
-import * as combinedCourseBlueprintRepository from './combined-course-blueprint-repository.js';
+import * as combinedCourseBlueprintRepository from './combined-course-blueprints/combined-course-blueprint-repository.js';
 
 export const findByOrganizationId = async ({ organizationId, locale }) => {
   const [blueprints, sharedTargetProfiles] = await Promise.all([
