+++ api/src/quest/infrastructure/repositories/eligibility-repository.js
@@ -1,6 +1,6 @@
 import { Eligibility } from '../../domain/models/quests/aggregates/Eligibility.js';
 // We import this repository here to avoid calling it in both dependencies and repositories in ./index.js
-import * as questOrganizationLearnerParticipationRepository from './organization-learner-participation-repository.js';
+import * as questOrganizationLearnerParticipationRepository from './combined-course-participations/organization-learner-participation-repository.js';
 export const find = async ({ userId, organizationLearnerWithParticipationApi }) => {
   const result = await organizationLearnerWithParticipationApi.find({ userIds: [userId] });
   return result.map(toDomain);
