+++ api/tests/prescription/organization-learner/integration/domain/usecases/get-organization-to-join_test.js
@@ -2,7 +2,7 @@ import * as campaignRepository from '../../../../../../src/prescription/campaign
 import { OrganizationToJoin } from '../../../../../../src/prescription/organization-learner/domain/models/OrganizationToJoin.js';
 import { getOrganizationToJoin } from '../../../../../../src/prescription/organization-learner/domain/usecases/get-organization-to-join.js';
 import { repositories } from '../../../../../../src/prescription/organization-learner/infrastructure/repositories/index.js';
-import * as combinedCourseRepository from '../../../../../../src/quest/infrastructure/repositories/combined-course-repository.js';
+import * as combinedCourseRepository from '../../../../../../src/quest/infrastructure/repositories/combined-courses/combined-course-repository.js';
 import { expect } from '../../../../../test-helper.js';
 import { databaseBuilder } from '../../../../../tooling/databases.js';
