+++ api/tests/quest/integration/infrastructure/repositories/combined-course-participation-repository_test.js
@@ -6,7 +6,7 @@ import {
   OrganizationLearnerParticipationStatuses,
   OrganizationLearnerParticipationTypes,
 } from '../../../../../src/quest/domain/models/combined-course-participations/entities/OrganizationLearnerParticipation.js';
-import * as combinedCourseParticipationRepository from '../../../../../src/quest/infrastructure/repositories/combined-course-participation-repository.js';
+import * as combinedCourseParticipationRepository from '../../../../../src/quest/infrastructure/repositories/combined-course-participations/combined-course-participation-repository.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../tooling/databases.js';
