+++ api/tests/quest/unit/infrastructure/repositories/module-repository_test.js
@@ -1,7 +1,7 @@
 import sinon from 'sinon';
 
 import { Module } from '../../../../../src/quest/domain/models/combined-courses/entities/Module.js';
-import * as moduleRepository from '../../../../../src/quest/infrastructure/repositories/module-repository.js';
+import * as moduleRepository from '../../../../../src/quest/infrastructure/repositories/combined-courses/module-repository.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Repositories | Module Repository', function () {
